Restructure `DashboardPageSpeed` to fix ESLint error.
